enhancement - Added tests for CSP and fixed implementation.
Note that the test script cannot yet find the CSP function, since it is defined
in a private directory. An autosync between the corresponding private dirs
still needs to be made.

Please note that CSP is one convoluted mess. The original papers (Koles and
Ramoser) contain lots of notational mistakes, and tweaks that are irrelevant to
the optimization problem (ad hoc rescaling the covariance etc.). As such, I
have focused on a correct implementation of the core functionality. The ad-hoc
tweaks can in those papers can still be applied before the covariance matrices
are passed to the CSP function.


git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@5160 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0