cfg.bpfreq: resolve ambiguity in function help
changed the help [low high] to [lowFreq highFreq]. Otherwise it is
ambiguous, whether low high means the lowpass, highpass frequency (which
would in terms of absolute values be the same as [high low]) or whether
it rather means low frequency, high frequency (which in terms of the
filters would mean[highpass lowpass]).

Judging from the filter function, [lowFreq, highFreq] is correct.