Adopt to changes in OpenMEEG Version 2.4
The new version of OpenMEEG (2.4) has changed its output because it can now handle also nested topologies. The order of the variables in the system matrix has changed. The sensor interpolations done in ft_prepare_vol don't work anymore and lead to wrong results.
Using the sensor interpolations calculated by OpenMEEG itself solves this for all versions as it is always matchin the head matrix. Additionally it might bring qualitative improvements as it is directly derived from the head model.
You will need an additional function openmeeg_sensinterpolmat which I am happy to provide.
Note: The polarity of the voltages flip but this is now correct as it was wrong before (validated with FEM head models).