Handling of (often occuring) case that Physical Minimum larger than Maximum in header
the error of Physical Minimum larger than Maximum often occurs, 
instead of replacing the Physical Minimum and Maximum with the Digital Minimum larger than Maximum,
the code tries to correct this common error in the header by swiching the Minimums and Maximums in the respective channels.
This handling is not perfect, but under most circumstances leads to the correct signals as they also apear (exaclty) in the EDFbrowser with the correct scaling and polarity of the signals. This handling avoids the more severe case that the scaling of ALL signals are falsely put to the Digital Minimum and Maximum.