Proposed Handling of EDF.PhysMin >= EDF.PhysMax, and potential failing of the quick read_16bit
The case "EDF.PhysMin >= EDF.PhysMax" that creates a warning typically occurrs if the signal shall be displayed reverse (negative up, positive down) in a viewer (e.g. EDFbrowser). That means that at least one channel has this property. 

Currently if this is the case for even one channel it affects the handling of all other channels by using the Digital Maximum and Minimum instead, which creates wrong signal values for all channels. Also the signal is falsely inverted (188ff).

Proposed here is that the Physical minimum and maximum are keept in this case leading to the correct, calibrated (scaled) signals with the correct signal offsets (that is negative values in the file stay negative values in the output, and positive values in the file stay positive in the output, irrespective how it shall be displayed in a viewer). Since the user might still expect the signal to be reversed as compared to another viewer (as might be the case for external EDF viewers like EDFbrowser) the user gets a warning that includes a lists with the channels in question to check their polarity and scaling again.

Additional suggestion: as sometimes the quick read in of read_16bit fails, then still the slow method can be tried.