Update ft_multiplotER.m
Currently cfg.linestyle option is meaningless. In the old version on line 478 , cfg.linestyle{i} does not make sense because the line is not even looping around i which should be from 1: Ndata. Proposing the following changes so ft_multiplotER can plot lines of different style (dashed, dotted etc.). I tested the changes and it works. This same change must be done on ft_singleplotER as well, which I will fork/and edit.