Update ft_singleplotER.m
This is a follow up change to have a meaningful cfg.linestyle option. It was nonfunctional both in ft_multiplotER and ft_singleplotER. I edited the ft_multiplotER and submited a Pull request separately. Upon plotting using function ft_multiplotER, in the interactive model when clicking a particular channel ft_singleplotER is called. The changes proposed are essential so the chosen channel reflects the linestyles using in multiplotER.