left mouse botton is not used for selecting electrode positions anymore
Adds the functionality to select the electrode position with only the middle mouse button (can be modified for using only the right mouse button, there is a comment mentioning that) . Instead of accidently clicking the left mouse button because one may want to use the matlab 3D rotate functionality. Other keys and their functionality remain the same.