on fail, read_brainvision_vhdr.m now looks for .eeg and .vmrk files with same name as .vhdr (#1100)
* Update read_brainvision_vhdr.m

Currently FieldTrip looks to the .vhdr file to determine the name of the associated .vmrk and .eeg files.  This causes problems when a researcher renames the files without realizing that .vhdr remembers the original names and will still look for them.  This revision leaves this behavior intact but adds the additional functionality of looking for .vmrk and .eeg files with the same name as the .vhdr file when the ones named in the .vhdr file cannot be found.  It also issues a note in the Command Window when a .vmrk or .eeg file named in the .vhdr file differs from the current name of the .vhdr file.

* Update read_brainvision_vhdr.m