Update in ft_redefinetrial to enable overlap
So far the option allow for overlap is not possible because the option is not retrieved and not taken into account by ft_fetch_data. With this modification, adding cfg.overlap = true will enable it when calling ft_redefintrial. Apparently this was to prevent some issues, but when you need to manually add some buffer that will be further removed, for instance, then it can become useful to have data overlap.