Update ft_prepare_sourcemodel.m
As requested, I implemented cfg.mode option for 'basedon', of which there can only be one. Now the question is whether you'd want this option to be specified by the user or determine it automatically? Since ft_prepare_sourcemodel is not necessarily called directly by the user, maybe the latter is better?  I named the option cfg.mode instead of cfg.method for if you want it to be user-specified.(cfg.method can run into conflicts if ft_prepare_sourcemodel is for example called by ft_sourceanalysis).