Fix #1361 (#1362)
* Fix #1361

if cfg.sourcemodel is based on a file, it is not mutually exclusive with the other options for constructing a sourcemodel.

* Update ft_prepare_sourcemodel.m

* Update ft_prepare_sourcemodel.m

As requested, I implemented cfg.mode option for 'basedon', of which there can only be one. Now the question is whether you'd want this option to be specified by the user or determine it automatically? Since ft_prepare_sourcemodel is not necessarily called directly by the user, maybe the latter is better?  I named the option cfg.mode instead of cfg.method for if you want it to be user-specified.(cfg.method can run into conflicts if ft_prepare_sourcemodel is for example called by ft_sourceanalysis).

* Update ft_prepare_sourcemodel.m

* Update ft_prepare_sourcemodel.m

Replaced one more 'mode' by 'method'

* Update ft_prepare_sourcemodel.m

Co-Authored-By: Mats <m.vanes@donders.ru.nl>

Co-authored-by: Robert Oostenveld <r.oostenveld@gmail.com>
Co-authored-by: Jan-Mathijs Schoffelen <j.schoffelen@donders.ru.nl>