renamed master/slave into controller/worker
The recent discussions related to blacklivesmatter made me realize that it is not OK to perpetuate the use of the negative words "master" and "slave" for https://en.wikipedia.org/wiki/Master/slave_(technology). In this specific case it is not a problem to use controller/worker instead.
I searched and replaced all occurences in the code, and renamed some files. I deleted the mex files, these should be recompiled.