changed the internal logic for spm2/spm8/spm12 and old/new/mars
the complex configuration of conditional "if" statements  could be simplified quite a bit since old applies to all three 2/8/12 and new and mars only apply to 12.

I also moved all of the undocumented options further up in the help to make them visible. I think that we expect people to use those anyway, so better be explicit about them. Also, these are not new/experimental options, but have been around for a long time and unlikely to disappear soon.

Furthermore, I tightened the input configuration checks a bit.