ENH - deprecated cfg.warpmni, since cfg.method is more clear
cfg.warpmni was used in combination with cfg.mri to determine whether to spatially warp from an MNI template, or whether to restrict sources to gray matter (from an MRI segmentation on the fly). This change in cfg-parsing makes it more clear, and I removed cfg.warpmni from the help.