Update ft_prepare_montage.m
New feature: creation of a montage for Laplacian re-referencing. This feature is particularly useful for iEEG data; each channel is re-referenced against the average of its two closest neighbors within the same electrode (the top and bottom channels of the electrode are re-referenced against their only closest neighbor). Note: the current version of this feature does not automatically separate the electrodes during the processing. It is recommended to perform the re-referencing separately from the rest of the preprocessing and to re-reference each group of channels  separately.