data2bids onset duration added to events.tsv
Some forms of cfg.events input are of form "onset" "duration" whereas others are of form "begsample", "endsample", "offset". In the later case, an error was given when the table gets sorted so that onset and duration are the first columns of the table (old lines 1590 to 1603). It should now be solved by the addition of columns "onset" and ""duration" to tables of the latter form.