improve formatting for printing neighbourhood structure
I made the transposing of columns and rows optional, as well ass removing the last newline. The printmat function is now more similar to the printcell function.
When a cell-array or a numeric array is a row (i.e. 1xN), it is now printed more compactly and not spread over multiple lines.