always assign the trigger value just _after_ it changed
This was different for the up and down cases, which in my eyes does not make sense. I made the code consistent so that the value that is assigned is always the value after the change is detected. If you want the value before, you should specify a trigshift of -1