fixed recursion error for neuromag306alt
I changed the behaviour for neuromag306/neuromag306alt and neuromag122/neuromag122alt: it used to return a cell-array with both regular and alternative labels, now it returns regular labels, or alternative labels, but not both combined.
I also checked with ft_read_header what the regular/alternative is (with/without space), regretfully it is just the opposite for 122/306.