Issue #1841 (#1846)
* New sphsplint.m file

This file aims to replace both splint (called only by ft_scalpcurrentdensity) and sphericalSplineInterpolate (called only by ft_channelrepair). It is a faster, more readable version of splint, returning the interpolation filters for the potential and the surface Laplacian, instead of the values themselves.

* New mplgndr.m file

Function to calculate the associated Legendre functions. Two differences with plgndr:
1. Works with matrices instead of scalars.
2. Returns the associated functions of order K and degrees 0 to N in a single call.

As the calculation of the i-th degree function is iterative (uses the i-1-th and i-2-th degree functions), and many functions use summation over degrees, this implementation saves a lot of computation time.

* Test file for issue #1841

Tests splint and sphericalSplineInterpolate to check weather they provide the same result.
Also compares this result with the new function sphsplint designed to replace both.

* Update test_issue1841.m

* Adding call to sphsplint

Replaces the call to sphericalSplineInterpolate by a call to sphsplint.

* Adding call to sphsplint

Replaces the call to splint by a call to sphsplint.
Calls sphsplint only once, and applies the filter to each trial.
Also adds a check for the spherical fit.

* Typo

* Adding CD to private folder

* Only one electrode set as imput

Adding the option of providing only one electrode set as input.
Useful for the calculation of the Laplacian if no bad channels are provided.

* Update ft_scalpcurrentdensity.m

* Changing fake-tab from 4 spaces to 2.

* Changing fake-tab from 4 spaces to 2

* Using ft_apply_montage with spherical splines

Uses ft_apply_montage to apply the spatial filter generated using spherical splines.

Also removes the unnecessary code re-filling the metadata in the output with that in the input (and messing with the electrode labels, see https://github.com/fieldtrip/fieldtrip/issues/1841#issuecomment-893580074 )

Co-authored-by: Jan-Mathijs Schoffelen <1517611+schoffelen@users.noreply.github.com>