moved failing test script out of the way
the error is due to the two trials in the data not overlapping, causing the variance  to become nan and the plot of the error bars to fail. I believe that the new code is an improvement of the old code, and that this test script therefore needs to be revised.