networkanalysis: use parfor instead of for
With some small adaptations, it is possible to use a parfor loop in
ft_networkanalysis. This potentially speeds up the calculation, when the
parallel computing toolbox is available.

Possibly, this could be further sped up by converting the nested loops into
a single parfor loop. However, this requires a good benchmark testcase to be
available, to test the difference.