timelockanalysis: refactor parfor loop to use less memory
For the not-keeptrials case, it is possible to use less memory, by adding results
within the loop. This was done with the regular for loop. Rework the parfor
implementation to also do the reduction within the loop. Does require having two
implementations of the parfor loop, because variables need to always be assigned to
in the same way within a single parfor loop.