FIX - don't start with spm_affreg or spm_normalise when the coordinates already match
There was a bug for input MRI data that would always be converted to MNI using SPM code. I detected this when doing ft_sourceplot with cfg.atlas containing a segmented MRI in CTF coordinates.

I also made some updates to the order of the initial default settings and input handling, that should not have functional consequences. Furthermore, I fixed the feedback plot.