See #2301 - give suggested tests when PR is submitted
This uses a github action and python script to add a comment to the
pull request with the suggested test scripts. The suggestion is
based on seacrhing for "DEPENDENCY" on any of the changed files.
Furthermore, it splits them in internal and external tests based
on the "DATA" requirement of the test scripts.

It is probably not yet optimal in the sense that it will rerun upon
every update to the PR.