merge 'spinning', 'gareth', and 'moiseev' cases
The three fixedori-approaches 'spinning', 'gareth', and 'moiseev' all
essentially perform the same procedure for orientation estimation. They
only differ in whether the whole orientation or only its tangential
component are estimated, and in how the noise covariance matrix is
estimated. Hence they can be merged into a single approach. Furthermore,
the tangential orientation in the 'spinning' case is now estimated from the
leadfield, instead of by fitting a sphere to the headmodel.